backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyMySQL==1.1.2
cryptography==44.0.0
django-filter==25.2
drf-spectacular==0.28.0
requests==2.32.3
django-cors-headers==4.6.0
